Apply AM vowel check to all Lao content
